fix: Scaling bounds should parse float not int (#10026)
* fix: Scaling bounds should parse float not int

Currently the API can do this, the model supports this, but the parseInt removes decimals from the UI

* fix: Decimal support for upper/lower bound ASG tests

+++ packages/amazon/src/serverGroup/details/scalingPolicy/upsert/step/StepPolicyAction.tsx
@@ -107,7 +107,7 @@ export const StepPolicyAction = ({
                   value={step.metricIntervalLowerBound}
                   max={step.metricIntervalUpperBound}
                   onChange={(e) =>
-                    updateStep({ ...step, metricIntervalLowerBound: Number.parseInt(e.target.value) }, index)
+                    updateStep({ ...step, metricIntervalLowerBound: Number.parseFloat(e.target.value) }, index)
                   }
                   inputClassName="action-input"
                 />
@@ -122,7 +122,7 @@ export const StepPolicyAction = ({
                   value={step.metricIntervalUpperBound}
                   min={step.metricIntervalLowerBound}
                   onChange={(e) =>
-                    updateStep({ ...step, metricIntervalUpperBound: Number.parseInt(e.target.value) }, index)
+                    updateStep({ ...step, metricIntervalUpperBound: Number.parseFloat(e.target.value) }, index)
                   }
                   inputClassName="action-input"
                 />
